\name{palyoplot_get2ndAxis}
\alias{palyoplot_get2ndAxis}
\title{Create 2nd y-axis for plotted data}
\description{Creates a secondary y-axis (ex: depth to go along with age). For use with palyoplot_plotData}
\usage{palyoplot_get2ndAxis(
interval,
ages,
agemodel = TRUE,
top = NULL,
bot = NULL,
ages.col.secondary = 1,
ages.col.age = 5
)}
\arguments{
  \item{interval}{at what interval should the secondary axis be made (ex: 10)}
  \item{ages}{a matrix of y-axis data (ex: depths, ages)}
  \item{agemodel}{is "ages" an age model (for instance, as generated by Bacon). TRUE. If a data frame of 2 varables, then FALSE. Default TRUE}
  \item{top}{top y-axis value (limit the dataset). Default NULL}
  \item{bot}{bottom y-axis value (limit the dataset). Default NULL}
  \item{ages.col.secondary}{index of secondary axis data. Default 1}
  \item{ages.col.age}{index of primary axis data (ages). Default 5}
}
\value{
  \item{y2at}{location along the primary axis where the secondar labels should be placed}
  \item{y2labels}{secondary axis labels}
}
\author{Anna Klimaszewski-Patterson}
\seealso{
  \code{\link{palyoplot_plotTaxa}}
}
\examples{
## EXAMPLE 1: generic data
botAge = 100; #max age to plot
topAge = -60; #min age to plot
Ages = data.frame(depth=c(0,20,40,60), age=c(-60,-10,40,100))
xdata = data.frame(Abies=c(50,12,24,32),
                   Pinus=c(54,20,44,19),
                   Quercus=c(12,50,20,11),
                   Roseaceae=c(4,9,12,2))

#create secondary y-axis information (2 y-axes)
y2 = palyoplot_get2ndAxis(interval=10, top=topAge, bot=botAge,
                          ages=Ages, ages.col.secondary=1, ages.col.age=2,
                          agemodel=FALSE)

#plot and return ggplot2 object
g = palyoplot_plotTaxa(x=xdata, y=Ages[,2],
                       y.label="cal yr BP", ylim=c(topAge, botAge),
                       y.break=50, y2=y2, y2.label="depth (cm)",
                       x.label="Percentages")
### EXAMPLE 2: using package data
maxAge = 115;
minAge = -20;

y2 = palyoplot_get2ndAxis(interval=2, top=minAge, bot=maxAge,
                          ages=pp_agemodel, ages.col.secondary=1, ages.col.age=5,
                          agemodel=TRUE)

g = palyoplot_plotTaxa(x=pp_xdata, y=pp_ydata,
                       y2=y2, y2.label = "depth (cm)",
                       y.label="cal yr BP", x.label="Percentages")

}
