\name{pp_agemodel}
\alias{pp_agemodel}
\docType{data}
\title{Palyoplot sample age model (data type: list)}
\description{Age model to use with your dataset. For example, the output from Bacon or Clam (depth and age at a mninimum)}
\usage{data("pp_agemodel")}
\format{
  A data frame with 496 observations on the following 7 variables.
  \describe{
    \item{\code{depth}}{a numeric vector}
    \item{\code{min}}{a numeric vector}
    \item{\code{max}}{a numeric vector}
    \item{\code{median}}{a numeric vector}
    \item{\code{wmean}}{a numeric vector}
    \item{\code{accrate}}{a factor with levels \code{10.5} \code{10.6} \code{10.7} \code{10.9} \code{11.0} \code{12.9} \code{13.0} \code{13.1} \code{13.2} \code{13.3} \code{13.4} \code{13.5} \code{13.6} \code{13.7} \code{13.8} \code{13.9} \code{14.0} \code{14.1} \code{14.2} \code{14.3} \code{14.4} \code{14.5} \code{14.6} \code{14.7} \code{14.8} \code{14.9} \code{15.0} \code{15.1} \code{15.2} \code{15.3} \code{15.4} \code{15.5} \code{15.7} \code{15.8} \code{15.9} \code{16.2} \code{16.8} \code{17.3} \code{17.4} \code{17.5} \code{17.6} \code{17.7} \code{17.8} \code{18.1} \code{18.4} \code{19.0} \code{19.3} \code{19.6} \code{19.7} \code{19.8} \code{19.9} \code{20.1} \code{20.2} \code{20.3} \code{20.4} \code{20.5} \code{20.6} \code{20.7} \code{20.8} \code{20.9} \code{21.0} \code{22.5} \code{22.7} \code{22.8} \code{23.1} \code{23.4} \code{23.5} \code{23.6} \code{23.7} \code{23.8} \code{23.9} \code{24.0} \code{24.1} \code{24.2} \code{24.3} \code{24.6} \code{24.7} \code{24.9} \code{25.0} \code{25.1} \code{25.2} \code{25.3} \code{25.6} \code{26.0} \code{26.2} \code{26.3} \code{26.5} \code{26.7} \code{27.1} \code{27.2} \code{28.1} \code{28.3} \code{28.4} \code{28.6} \code{28.7} \code{29.0} \code{29.2} \code{29.4} \code{29.6} \code{29.7} \code{29.8} \code{29.9} \code{30.3} \code{30.4} \code{30.6} \code{30.8} \code{31.0} \code{31.2} \code{31.3} \code{31.4} \code{31.6} \code{31.7} \code{31.8} \code{32.1} \code{32.5} \code{32.8} \code{32.9} \code{33.0} \code{33.1} \code{33.2} \code{33.3} \code{35.8} \code{36.1} \code{36.5} \code{36.6} \code{36.7} \code{39.4} \code{39.9} \code{40.1} \code{40.2} \code{40.3} \code{40.4} \code{40.7} \code{40.8} \code{40.9} \code{41.0} \code{41.4} \code{41.5} \code{42.5} \code{42.6} \code{42.8} \code{43.1} \code{44.0} \code{44.1} \code{44.4} \code{44.6} \code{44.7} \code{44.8} \code{44.9} \code{45.0} \code{45.1} \code{45.2} \code{45.3} \code{45.4} \code{45.5} \code{45.6} \code{45.8} \code{45.9} \code{46.2} \code{46.3} \code{47.1} \code{48.1} \code{48.2} \code{51.9} \code{52.1} \code{52.2} \code{52.3} \code{52.4} \code{53.1} \code{53.7} \code{54.2} \code{54.4} \code{54.5} \code{54.7} \code{55.1} \code{55.2} \code{56.5} \code{56.6} \code{56.7} \code{56.8} \code{56.9} \code{57.2} \code{57.3} \code{57.4} \code{57.5} \code{57.6} \code{57.8} \code{58.2} \code{58.4} \code{58.9} \code{59.1} \code{59.2} \code{59.4} \code{59.6} \code{59.7} \code{59.9} \code{6.3} \code{6.4} \code{6.5} \code{6.6} \code{6.7} \code{6.8} \code{6.9} \code{60.1} \code{60.3} \code{61.7} \code{61.8} \code{62.0} \code{62.1} \code{62.7} \code{63.5} \code{63.7} \code{63.9} \code{64.0} \code{64.2} \code{64.3} \code{64.6} \code{65.9} \code{66.0} \code{66.2} \code{66.3} \code{67.1} \code{7.0} \code{7.1} \code{7.3} \code{73.0} \code{73.5} \code{73.7} \code{73.8} \code{NULL}}
    \item{\code{id}}{a numeric vector}
  }
}
\examples{
data(pp_agemodel)
}
\keyword{datasets}
